inventory: add network tag as alternative to vlan
Each interface must be in a named network (and gets assigned to
'default', if missing). Each vlan is a named network - but each
network not necessarily is a vlan, there are various other ways of
isolating them as well.

Therefore in some cases just calling it 'network' is more sensible
than 'vlan'.