feat: Register Python version on Linux systems
Add tasks to register the Python version on Linux systems. This
includes a shell command to retrieve the Python version and set
it as a fact named `python_version`. The first task uses `python3
--version` piped through `awk` to extract the major and minor
version components. The second task sets the `python_version`
fact if the extracted version is defined. Both tasks apply to
Linux systems and are tagged with `base_config` and `access_setup`.

The purpose of these changes is to readily access the Python
version information for configurations or decision-making in
subsequent playbook tasks.