feat: Add oscdimg support for ISO creation
Updated the script to support both oscdimg and imgburn for ISO
creation. Added a check for oscdimg's path and adjusted the logic to
use it if available, falling back to imgburn otherwise. The changes
include copying 'etfsboot.com' and 'efisys.bin' when using oscdimg
as the ISO creation tool.

This allows creation of images supporting both BIOS and EFI boot, and
in most cases allows image creation without installing extra software.